fixed from PVS-Studio
V668 There is no sense in testing the 'Vector' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. vector.h 59
V668 There is no sense in testing the 'newindex' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. vector.h 1230
V668 There is no sense in testing the 'section' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. ini.cpp 118
V668 There is no sense in testing the 'entry' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. ini.cpp 150